Add emotion chip and admin functions (#32)
Add categories, a set of values to respond with in certain situtations.
This allows the bot to respond with a random response from a
configurable list. The admin functions allow adding and remove items
from the categories. Add emotion chip as a simple demonstration of
categories, along with a few response to make the bot a little more
friendly (as well as responding to hostility!).